Update JacoComm.m
Adapting the connect method to support new functions. I add new arguments when calling the JacoMexInterface function. I also add a call to the function setNumJoints to detect if the arm is a 4. 6 or 7 DOF
Adapting the disconnect method to support new functions. I add new arguments when calling the JacoMexInterface function
Adding the setNumJoints method