Update Kinova API (#1)
* Update JacoSDKWrapper.h

Changing numJoint from 6 to 7

* Update JacoSDKWrapper.h


getting back to numJoint = 6

* Update JacoSDKWrapper.h

Adding New indices in class FunctionIndices for the new functions. New indices start at 18 and end at 28.

* Update JacoSDKWrapper.h


adding function header for 
-SendCartesianPosition 
-SendCartesianVelocity 
-GetDOF 
-StartForceControl 
-StopForceControl 
-GetEndEffectorOffset 
-SetEndEffectorOffset 
-GetProtectionZone 
-EraseProtectionZones
-SetProtectionZone 
-GetGlobalTrajectoryInfo

* Update JacoSDKWrapper.cpp

Adding function pointer for new functions

* Update JacoSDKWrapper.cpp

Loading the new functions from the librairy

* Update JacoSDKWrapper.cpp

Changing the function getJointPosition, getJointVelocity, getJointTorque and getJointTemperature to make it returns the value for the seventh actuator

* Update JacoSDKWrapper.cpp

change the functions sendJointPositions, sendJointAndFingerPositions, sendJointVelocities, sendJointTorques andsendFingerPositions to make it works with a 7DOF.

* Update JacoSDKWrapper.cpp

Adding the definitions of the new functions, 
 -SendCartesianPosition 
-SendCartesianVelocity 
-GetDOF
-StartForceControl 
-StopForceControl 
-GetEndEffectorOffset 
-SetEndEffectorOffset 
-GetProtectionZone 
-EraseProtectionZones 
-SetProtectionZone 
-GetGlobalTrajectoryInfo

* Update JacoMexInterface.cpp

Adding the new inputs pointers for the new functions
Adding the new outputs pointers for the new functions
Changing the number of input and output in the verification for proper number of arguments
Changing the verification on inputs are real to make it verify all the inputs
Adding individual verification of the new inputs (CartPos, CartVel, Offset, Zone)
Create pointers to the real data in the input matrices for the new pointers
Create the output for the outputs pointers
Adding the new cases for the new functions

* Update compileMexInterface.m

Adding a change directory to make sure the new JacoMexInterface.mexw64 file always overwrite the old one
Changing path for

* Update MexFunctionIDs.m

Adding the IDs for the new functions (18 to 28)

* Update JacoCommandModes.m

Adding new modes to cartesian control

* Update JacoComm.m

Changing NumJoints from constant private property to variable private property
Adding CartParam as a constant private Property
Adding CartPosCmd, CartVelCmd, OffsetCmd, ZoneCmd as privates variables properties
Adding DOF, EndEffectorOffset, ProtectionZone, TrajectoryInfo as protected variables properties

* Update JacoComm.m

Adapting the connect method to support new functions. I add new arguments when calling the JacoMexInterface function. I also add a call to the function setNumJoints to detect if the arm is a 4. 6 or 7 DOF
Adapting the disconnect method to support new functions. I add new arguments when calling the JacoMexInterface function
Adding the setNumJoints method

* Update JacoComm.m

Adapting each time we call JacoMexInterface to pass all the new property and get all the new informations from the new functions

* Update JacoComm.m

Adding the publics and privates methods SendCartesianPositionCommand, SendCartesianVelocityCommand,
setEndEffectorOffset and setProtectionZone.

* Update JacoComm.m

Adding publics and privates methods to get these properties DOF, EndEffectorOffset, ProtectionZone, TrajectoryInfo

* Update JacoComm.m

Adding methods StartForceControl, StopForceControl and EraseAllProtectionZone

* Update JacoComm.m

Adapting the simulink methods to call the new functions

* update JacoSDKWrapper.h

change NUM_JOINTS from 6 to 7 to support 7 kinova's 7 DOF arm

* adjustment

Adding pointer to the function that get the cartesian command in JacoSDKWrapper.cpp

Changing the path in compileMexInteface

Change the num joints property to the number 7 because we switch this property from constant to variable earlier.

create an automatic way to add path to matlab in testJacoComm

* update JacoMexInterface

Adapt the condition of the verification of jntPos, jntVel and jntTorque to support 4,6 and 7 DOF

* Adding exemples to testJacoComm

Adding the new functions to the testJacoComm

Remove print screen for debug

* Minor bug fixes to make the simulink run

* Adding the cartesian control exemple to simulink

* Adding a new guide about how to add other API's functions

* Adding the simulink in 2017 version

* Publishing R2017b version of the simulink file

* GetDof Function wotk better now

* Adding JacoViewer

* bug fixed

* Final change for supporting 6 and 7 DOF

* delete useless files